Debugging tests.
