+++ src/example-streaming/Program.cs
@@ -3,7 +3,7 @@ using QuestDB.Ingress;
 
 var rowsToSend = 1e6;
 
-using var sender = new Sender("http::addr=localhost:9000;auto_flush=on;auto_flush_rows=75000;");
+using var sender = new SenderOld("http::addr=localhost:9000;auto_flush=on;auto_flush_rows=75000;");
 
 var timer = new Stopwatch();
 timer.Start();
