+++ src/net-questdb-client-tests/TcpTests.cs
@@ -32,7 +32,7 @@ using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
 using QuestDB.Ingress;
 using QuestDB.Ingress.Enums;
-using QuestDB.Ingress.Misc;
+using QuestDB.Ingress.Utils;
 
 namespace net_questdb_client_tests;
 
@@ -47,7 +47,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var ls = new Sender($"tcp::addr={_host}:{_port};");
+        using var ls = new SenderOld($"tcp::addr={_host}:{_port};");
         ls.Table("metric name")
             .Symbol("t a g", "v alu, e")
             .Column("number", 10)
@@ -69,7 +69,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var ls =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};username=testUser1;token=NgdiOWDoQNUP18WOnb1xkkEG5TzPYMda5SiUOvT1K0U=;");
 
         ls.Table("metric name")
@@ -92,7 +92,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         Assert.That(
-            () => new Sender($"tcp::addr={_host}:{_port};username=invalid;token=foo=;")
+            () => new SenderOld($"tcp::addr={_host}:{_port};username=invalid;token=foo=;")
             ,
             Throws.TypeOf<AggregateException>().With.InnerException.TypeOf<IngressError>().With.Message
                 .Contains("Authentication failed")
@@ -110,7 +110,7 @@ public class TcpTests
         Assert.That(
             async () =>
             {
-                using var sender = new Sender(
+                using var sender = new SenderOld(
                     $"tcp::addr={_host}:{_port};username=testUser1;token=ZOvHHNQBGvZuiCLt7CmWt0tTlsnjm9F3O3C749wGT_M=;");
 
                 for (var i = 0; i < 10; i++)
@@ -173,7 +173,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender($"tcp::addr={_host}:{_port};init_buf_size=2048;");
+            new SenderOld($"tcp::addr={_host}:{_port};init_buf_size=2048;");
         var lineCount = 500;
         var expected =
             "table\\ name,t\\ a\\ g=v\\ alu\\,\\ e number=10i,db\\ l=123.12,string=\" -=\\\"\",при\\ вед=\"медвед\" 1000000000\n";
@@ -203,7 +203,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender($"tcp::addr={_host}:{_port};init_buf_size=2048;");
+            new SenderOld($"tcp::addr={_host}:{_port};init_buf_size=2048;");
         var lineCount = 500;
         var expected =
             "table\\ name,t\\ a\\ g=v\\ alu\\,\\ e number=10i,db\\ l=123.12,string=\" -=\\\"\",при\\ вед=\"медвед\" 1000000000\n";
@@ -247,7 +247,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender($"tcp::addr={_host}:{_port};init_buf_size=2048;");
+            new SenderOld($"tcp::addr={_host}:{_port};init_buf_size=2048;");
         var lineCount = 500;
         var expected =
             "table\\ name,t\\ a\\ g=v\\ alu\\,\\ e number=10i,db\\ l=123.12,string=\" -=\\\"\",при\\ вед=\"медвед\" 1000000000\n";
@@ -292,7 +292,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender($"tcp::addr={_host}:{_port};init_buf_size=2048;tls_verify=unsafe_off;");
+            new SenderOld($"tcp::addr={_host}:{_port};init_buf_size=2048;tls_verify=unsafe_off;");
 
         var lineCount = 500;
         var expected =
@@ -330,7 +330,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var ls = new Sender($"tcp::addr={_host}:{_port};");
+        using var ls = new SenderOld($"tcp::addr={_host}:{_port};");
 
         ls.Table("neg name")
             .Column("number1", long.MinValue + 1)
@@ -351,7 +351,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var ls = new Sender($"tcp::addr={_host}:{_port};");
+        using var ls = new SenderOld($"tcp::addr={_host}:{_port};");
 
         ls.Table("doubles")
             .Column("d0", 0.0)
@@ -376,7 +376,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var ls = new Sender($"tcp::addr={_host}:{_port};");
+        using var ls = new SenderOld($"tcp::addr={_host}:{_port};");
 
         var ts = new DateTime(2022, 2, 24);
         ls.Table("name")
@@ -400,7 +400,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcps::addr={_host}:{_port};username=testUser1;token=NgdiOWDoQNUP18WOnb1xkkEG5TzPYMda5SiUOvT1K0U=;tls_verify=unsafe_off;");
 
         sender.Table("table_name")
@@ -424,7 +424,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var sender = new Sender($"tcp::addr={_host}:{_port};");
+        using var sender = new SenderOld($"tcp::addr={_host}:{_port};");
         string? nullString = null;
 
         Assert.That(
@@ -477,7 +477,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var sender_lim_127 = new Sender($"tcp::addr={_host}:{_port};");
+        using var sender_lim_127 = new SenderOld($"tcp::addr={_host}:{_port};");
         string? nullString = null;
 
         Assert.Throws<IngressError>(() => sender_lim_127.Table("abc\\slash"));
@@ -489,7 +489,7 @@ public class TcpTests
         Assert.Throws<IngressError>(() => sender_lim_127.Table("asdf\rsdf"));
         Assert.Throws<IngressError>(() => sender_lim_127.Table("asdfsdf."));
 
-        using var sender_lim_4 = new Sender($"tcp::addr={_host}:{_port};max_name_len=4;");
+        using var sender_lim_4 = new SenderOld($"tcp::addr={_host}:{_port};max_name_len=4;");
         Assert.Throws<IngressError>(() => sender_lim_4.Table("asffdfasdf"));
 
         sender_lim_127.Table("abcd.csv");
@@ -527,7 +527,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var sender = new Sender($"tcp::addr={_host}:{_port};");
+        using var sender = new SenderOld($"tcp::addr={_host}:{_port};");
         string? nullString = null;
 
         Assert.Throws<IngressError>(() => sender.Table(nullString));
@@ -551,7 +551,7 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
 
-        using var sender = new Sender($"tcp::addr={_host}:{_port};");
+        using var sender = new SenderOld($"tcp::addr={_host}:{_port};");
 
         sender.Table("good");
         sender.Symbol("asdf", "sdfad");
@@ -582,7 +582,7 @@ public class TcpTests
         var nowMillisecond = DateTime.Now.Millisecond;
         var metric = "metric_name" + nowMillisecond;
 
-        using var sender = new Sender($"tcp::addr={_host}:{_port};init_buf_size={256 * 1024};");
+        using var sender = new SenderOld($"tcp::addr={_host}:{_port};init_buf_size={256 * 1024};");
 
         for (var i = 0; i < 1E6; i++)
         {
@@ -612,7 +612,7 @@ public class TcpTests
         var metric = "metric_name" + nowMillisecond;
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};init_buf_size={64 * 1024};auto_flush=on;auto_flush_bytes={64 * 1024};");
 
         for (var i = 0; i < 1E6; i++)
@@ -632,7 +632,7 @@ public class TcpTests
     public async Task CannotConnect()
     {
         Assert.That(
-            () => new Sender($"tcp::addr={_host}:{_port};"),
+            () => new SenderOld($"tcp::addr={_host}:{_port};"),
             Throws.TypeOf<AggregateException>()
         );
     }
@@ -646,7 +646,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         Assert.That(
-            () => new Sender(
+            () => new SenderOld(
                 $"tcp::addr={_host}:{_port};init_buf_size=512;username=testUser1;token=NgdiOWDoQNUP18WOnb1xkkEG5TzPYMda5SiUOvT1K0U=;"),
             Throws.TypeOf<AggregateException>().With.InnerException.TypeOf<IngressError>().With.Message
                 .Contains("Buffer is too small to receive the message")
@@ -660,7 +660,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};");
         Assert.That(
             () => sender.Table("name")
@@ -677,7 +677,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};");
         sender.Table("neg name")
             .Column("привед", " мед\rве\n д")
@@ -695,7 +695,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};");
         Assert.Throws<IngressError>(
             () => sender.Table("name")
@@ -712,7 +712,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};");
         Assert.Throws<IngressError>(
             () => sender.Table("name")
@@ -729,7 +729,7 @@ public class TcpTests
         srv.AcceptAsync();
 
         using var sender =
-            new Sender(
+            new SenderOld(
                 $"tcp::addr={_host}:{_port};");
         Assert.Throws<IngressError>(
             () => sender.Column("number1", 123)
@@ -748,8 +748,8 @@ public class TcpTests
         using var srv = CreateTcpListener(_port);
         srv.AcceptAsync();
         Assert.AreEqual(
-            (await Sender.ConnectAsync("localhost", _port, protocol: ProtocolType.tcp)).Options.ToString(),
-            new Sender($"tcp::addr=localhost:{_port};init_buf_size=4096;").Options.ToString());
+            (await SenderOld.ConnectAsync("localhost", _port, protocol: ProtocolType.tcp)).Options.ToString(),
+            new SenderOld($"tcp::addr=localhost:{_port};init_buf_size=4096;").Options.ToString());
     }
 
     private static void WaitAssert(DummyIlpServer srv, string expected)
