--- src/net-questdb-client/Ingress/SenderFactory.cs
@@ -1,6 +1,0 @@
-namespace QuestDB.Ingress;
-
-public static class SenderFactory
-{
-    
-}
