+++ src/net-questdb-client/Ingress/Utils/Extensions.cs
@@ -1,0 +1,16 @@
+using QuestDB.Ingress.Senders;
+
+namespace QuestDB.Ingress.Utils;
+
+public static class Extensions
+{
+    public static void DisposeNullable<T>(this T? value) where T: IDisposable
+    {
+        if (value != null)
+        {
+            value.Dispose();
+        }
+    }
+
+    
+}
