+++ src/net-questdb-client/net-questdb-client.csproj
@@ -27,8 +27,4 @@
         <PackageReference Include="System.Net.Http" Version="4.3.4"/>
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Ingress\Contracts\" />
-    </ItemGroup>
-
 </Project>
